refactored shared functions into NFACT_base
